+++ static/app/components/asyncComponentSearchInput.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 import debounce from 'lodash/debounce';
 
 import {Client, ResponseMeta} from 'sentry/api';
-import Input from 'sentry/components/forms/controls/input';
+import Input from 'sentry/components/input';
 import LoadingIndicator from 'sentry/components/loadingIndicator';
 import {t} from 'sentry/locale';
