+++ static/app/components/modals/debugFileCustomRepository/http.tsx
@@ -4,9 +4,9 @@ import styled from '@emotion/styled';
 import {ModalRenderProps} from 'sentry/actionCreators/modal';
 import ActionButton from 'sentry/components/actions/button';
 import Button from 'sentry/components/button';
-import Input from 'sentry/components/forms/controls/input';
 import Field from 'sentry/components/forms/field';
 import SelectField from 'sentry/components/forms/selectField';
+import Input from 'sentry/components/input';
 import {
   DEBUG_SOURCE_CASINGS,
   DEBUG_SOURCE_LAYOUTS,
