+++ static/app/views/alerts/rules/metric/triggers/actionsPanel/actionSpecificTargetSelector.tsx
@@ -1,4 +1,4 @@
-import Input from 'sentry/components/forms/controls/input';
+import Input from 'sentry/components/input';
 import {t} from 'sentry/locale';
 import {Action, ActionType, TargetType} from 'sentry/views/alerts/rules/metric/types';
