+++ static/app/views/alerts/rules/metric/triggers/actionsPanel/actionTargetSelector.tsx
@@ -1,6 +1,6 @@
-import Input from 'sentry/components/forms/controls/input';
 import SelectControl from 'sentry/components/forms/selectControl';
 import TeamSelector from 'sentry/components/forms/teamSelector';
+import Input from 'sentry/components/input';
 import SelectMembers from 'sentry/components/selectMembers';
 import {Organization, Project, SelectValue} from 'sentry/types';
 import {
