+++ static/app/views/issueList/tagFilter.tsx
@@ -3,8 +3,8 @@ import debounce from 'lodash/debounce';
 
 import {addErrorMessage} from 'sentry/actionCreators/indicator';
 import {Client} from 'sentry/api';
-import Input from 'sentry/components/forms/controls/input';
 import SelectControl from 'sentry/components/forms/selectControl';
+import Input from 'sentry/components/input';
 import SidebarSection from 'sentry/components/sidebarSection';
 import {t, tct} from 'sentry/locale';
 import {Tag, TagValue} from 'sentry/types';
