+++ static/app/views/settings/organizationRelay/modals/form.tsx
@@ -1,10 +1,10 @@
 import styled from '@emotion/styled';
 
-import Input from 'sentry/components/forms/controls/input';
 import Textarea from 'sentry/components/forms/controls/textarea';
 import Field from 'sentry/components/forms/field';
 import FieldHelp from 'sentry/components/forms/field/fieldHelp';
 import TextCopyInput from 'sentry/components/forms/textCopyInput';
+import Input from 'sentry/components/input';
 import {t} from 'sentry/locale';
 import space from 'sentry/styles/space';
 import {Relay} from 'sentry/types';
